code refactoring 1차(sharing function, toolbar function) & search, picture edit function fix
