+++ NostelgiAlbum/HomeScreenViewController/SharingAlbumPush.swift
@@ -1,14 +1,13 @@
 import UIKit
 
 extension HomeScreenViewController {
-        
+    //HomeScreenViewController에서 ShareView를 push합니다.
     func pushShareView(path: URL) {
-        print("chomozzi")
         let shareVC = self.storyboard?.instantiateViewController(withIdentifier: "ShareViewController") as! ShareViewController
+        //.nost파일의 URL
         shareVC.filePath = path
+        //HomeScreenViewController의 collectionView Object
         shareVC.collectionViewInHome = collectionView
-        shareVC.modalPresentationStyle = .fullScreen
         self.navigationController?.pushViewController(shareVC, animated: true)
-        //self.present(shareVC, animated: true)
     }
 }
