+++ NostelgiAlbum/ImageFunc/ImageFunc.swift
@@ -1,7 +1,5 @@
 import Foundation
 import UIKit
-import Zip
-import RealmSwift
 
 func saveImageToDocumentDirectory(imageName: String, image: UIImage, AlbumCoverName: String) {
     // 1. 이미지를 저장할 경로를 설정해줘야함 - 도큐먼트 폴더,File 관련된건 Filemanager가 관리함(싱글톤 패턴)
@@ -73,45 +71,6 @@ func deleteImageFromDocumentDirectory(imageName: String) {
     }
 }
 
-func zipAlbumDirectory(AlbumCoverName: String) -> URL? {
-    guard let documentDirectory = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first else { return nil}
-    let dirURL = documentDirectory.appendingPathComponent(AlbumCoverName)
-    print("dirURL",dirURL.path)
-    let nostURL = documentDirectory.appendingPathComponent("\(AlbumCoverName).nost")
-    _ = FileManager.default
-    addAlbumInfo(coverData: AlbumCoverName)
-    do {
-        let zipfilePath = try Zip.quickZipFiles([dirURL], fileName: AlbumCoverName)
-        do {
-            if FileManager.default.fileExists(atPath: nostURL.path) {
-                do {
-                    try FileManager.default.removeItem(at: nostURL)
-                    print("nost파일 삭제 완료")
-                } catch {
-                    print("nost파일을 삭제하지 못했습니다.")
-                }
-            }
-            //try fileManager.moveItem(atPath: zipfilePath.path, toPath: nostURL.path)
-            if let data = try? Data(contentsOf: zipfilePath) {
-                let newURL = URL(filePath: nostURL.path)
-                try? data.write(to: newURL)
-                try FileManager.default.removeItem(at: zipfilePath)
-            } else {
-                print("Error rewrite file")
-            }
-        }
-    } catch {
-        print("Something went wrong")
-    }
-    
-    return nostURL
-    
-}
-
-func unzipAlbumDirectory(AlbumCoverName: String) {
-    
-}
-
 func resizeingImage(image: UIImage, width: Int, height: Int) -> UIImage? {
     let customImage = image
     let newImageRect = CGRect(x: 0, y: 0, width: width, height: height)
@@ -158,42 +117,4 @@ extension UIImage {
 
         }
     }
-}
-
-func addAlbumInfo(coverData: String) {
-    let realm = try! Realm()
-    let albumData = realm.objects(album.self).filter("AlbumTitle = '\(coverData)'")
-    let albumInfo = realm.objects(albumsInfo.self).filter("id = \(albumData.first!.index)")
-    var arrImageName = [String]()
-    var arrImageText = [String]()
-    var arrAlbumInfo = [String]()
-    
-    for i in 1...albumInfo.first!.numberOfPictures {
-        let albumImageName = albumData[i - 1].ImageName
-        let albumImageText = albumData[i - 1].ImageText
-        arrImageText.append("\(albumImageText)")
-        arrImageName.append("\(albumImageName)")
-    }
-    
-    arrAlbumInfo.append(String(albumInfo.first!.numberOfPictures))
-    arrAlbumInfo.append(albumInfo.first!.dateOfCreation)
-    
-    let imageNameInfo = arrImageName.joined(separator: "\n")
-    let imageTextInfo = arrImageText.joined(separator: "\n")
-    let albumInfoText = arrAlbumInfo.joined(separator: "\n")
-    
-    guard let documentDirectory = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first else { return }
-    let dirURL = documentDirectory.appendingPathComponent(coverData)
-    let nameInfoURL = dirURL.appendingPathComponent("\(coverData)imageNameInfo.txt")
-    let textInfoURL = dirURL.appendingPathComponent("\(coverData)imageTextInfo.txt")
-    let albumInfoURL = dirURL.appendingPathComponent("\(coverData)Info.txt")
-    
-    do {
-        try imageNameInfo.write(to: nameInfoURL, atomically: true, encoding: .utf8)
-        try imageTextInfo.write(to: textInfoURL, atomically: true, encoding: .utf8)
-        try albumInfoText.write(to: albumInfoURL, atomically: true, encoding: .utf8)
-    } catch {
-        print("Error writing file")
-    }
-    
 }
