+++ NostelgiAlbum/SceneDelegate.swift
@@ -8,34 +8,27 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         
         guard let _ = (scene as? UIWindowScene) else { return }
-//        let window = UIWindow(frame: UIScreen.main.bounds)
-//        let viewController = HomeScreenViewController()
-//        window.rootViewController = viewController
-//        //window.makeKeyAndVisible()
     }
-
+    /*
+     .nost파일을 통해 앱을 열 때 호출되는 함수
+     */
     func scene(_ scene: UIScene, openURLContexts URLContexts: Set<UIOpenURLContext>) {
-        //window?.rootViewController = HomeScreenViewController()
-//        let storyboard = UIStoryboard(name: "Main", bundle: nil)
-//        let vc : UIViewController = storyboard.instantiateViewController(withIdentifier: "HomeScreenViewController") as UIViewController
-//        window?.rootViewController = vc
-//        let url = URLContexts.first?.url
-//        if(url?.scheme == "file" && url?.pathExtension == "nost")
-//        {
-//            print("urlpath",url!.path)
-//            guard let homeScreenViewController = window?.rootViewController as? HomeScreenViewController else { return }
-//            homeScreenViewController.pushShareView(path: url!)
-//        }
+        //Main.storyboard set
         let storyboard = UIStoryboard(name: "Main", bundle: nil)
+        //Main.storyboard의 HomeScreenViewController객체를 가져옴
         let vc = storyboard.instantiateViewController(withIdentifier: "HomeScreenViewController")
+        //UINavigationController의 rootViewController를 HomeScreenViewController로 지정
         let navigationController = UINavigationController(rootViewController: vc)
+        //rootViewControlelr에 객체 값 할당
         window?.rootViewController = vc
         window?.rootViewController = navigationController
+        //NostelgiAlbum을 여는 경로가 .nost file일 때
         let url = URLContexts.first?.url
         if(url?.scheme == "file" && url?.pathExtension == "nost")
         {
-            print("urlpath",url!.path)
+            //navigationController.viewControllers.first == HomeScreenViewController
             guard let homeScreenViewController = navigationController.viewControllers.first as? HomeScreenViewController else { return }
+            //HomeScreenViewController의 pushShareView 동작(파일 URL을 같이 넣어줌)
             homeScreenViewController.pushShareView(path: url!)
         }
